% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogLikeliHood}
\alias{LogLikeliHood}
\title{LogLikeliHood}
\usage{
LogLikeliHood(param, datax, n_explanatory, n_transition,
  prob_initial_state_0, printLogLike)
}
\description{
\code{LogLikeliHood} is a convenience wrapper for function 'LogLikeliHood_FULL' that returns a single numeric scalar
with the log-likelihood of the Expectation-Maximization Maximum-Likelihood (EM-ML) algorithm as this is required by some
optimizers. This is a fast C++ implementation.
}
