% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogLikeliHood_byObs}
\alias{LogLikeliHood_byObs}
\title{LogLikeliHood_byObs}
\usage{
LogLikeliHood_byObs(param, datax, n_explanatory, n_transition,
  prob_initial_state_0, printLogLike)
}
\description{
\code{LogLikeliHood_byObs} is a convenience wrapper for function 'LogLikeliHood_FULL' that returns all observations'
log-likelihood constributions as required by certain optimizers.
}
