% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogLikeliHood_FULL}
\alias{LogLikeliHood_FULL}
\title{LogLikeliHood_FULL}
\usage{
LogLikeliHood_FULL(param, datax, n_explanatory, n_transition,
  prob_initial_state_0, printLogLike)
}
\description{
\code{LogLikeliHood_FULL} is a fast \code{C++} implementation that returns a numeric scalar \code{c(0)} with the log-likelihood of the
Expectation-Maximization Maximum-Likelihood (EM-ML) algorithm given observations for the explanatory variables governing direct
effects on the dependent variable, and observations for explanatory variables that govern the regime
transition process. It also returns various diagnostic statistics such as each individual observation's log-likelihood
contribution. If parameter 'printLogLike' = 1, the sum of the individual observation's log-likelihood contributions
is printed to the console. This fast \code{C++} implementation is several thousand times fast to the \R analog.
}
