% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLikeliHood_FULL_R.R
\name{LogLikeliHood_FULL_R}
\alias{LogLikeliHood_FULL_R}
\title{LogLikeliHood_FULL_R}
\usage{
LogLikeliHood_FULL_R(param, datax, n_explanatory, n_transition,
  prob_initial_state_0, printLogLike = 1)
}
\arguments{
\item{x, y}{two numeric values to be added up}
}
\value{
The addition of the inputs
}
\description{
\code{LogLikeliHood_FULL_R} returns a numeric scalar with the log-likelihood of the Expectation-Maximization
Maximum-Likelihood (EM-ML) algorithm given observations for the explanatory variables governing direct
effects on the the dependent variable and observations for explanatory variables that govern the regime
transition process. This implemenation is several thousand times slower compared to the fast C++ implementation.
    Author: Tobias Gummersbach
}
\section{Author}{

Tobias Gummersbach
}

\examples{
## add up two numbers
GoTobiGo(1,3)

## add up two values
x <- 1
y <- 3
GoTobiGo(x,y)
}
