% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogLikeliHood_min}
\alias{LogLikeliHood_min}
\title{LogLikeliHood_min}
\usage{
LogLikeliHood_min(param, datax, n_explanatory, n_transition,
  prob_initial_state_0, printLogLike)
}
\description{
\code{LogLikeliHood_min} is a convenience wrapper for function 'LogLikeliHood_FULL' that returns -1 * resulting log-likelihood
value as required by certain optimizers
}
