% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogLikeliHood_RunTimeCheck}
\alias{LogLikeliHood_RunTimeCheck}
\title{LogLikeliHood_RunTimeCheck}
\usage{
LogLikeliHood_RunTimeCheck(param, datax, n_explanatory, n_transition,
  prob_initial_state_0, printLogLike, n_it)
}
\description{
\code{LogLikeliHood_RunTimeCheck} is a convenience wrapper for function 'LogLikeliHood_FULL' that accepts the number of repetitive log-likelihood
calculations as an additional input. This is used to test the systems's computational speed.
value as required by certain optimizers
}
